update gallery, index

+++ app/Http/Controllers/VideoController.php
@@ -19,60 +19,47 @@ class VideoController extends Controller
     {
         $desa = Desa::find(1);
         $api_key = config('api.key');
-        Video::truncate();
-        $apiUrl = "https://www.googleapis.com/youtube/v3/search?";
-        $part = "part=snippet";
-        $channelId = "&channelId=$desa->channel_id";
-        $key = "&key=$api_key";
-        $maxResults = "&maxResults=50";
-        $nextPageToken = "&pageToken=";
-        $reload = true;
+        if ($api_key != "KOSONG") {
+            Video::truncate();
+            $apiUrl = "https://www.googleapis.com/youtube/v3/search?";
+            $part = "part=snippet";
+            $channelId = "&channelId={$desa->channel_id}";
+            $key = "&key={$api_key}";
+            $maxResults = "&maxResults=50";
+            $nextPageToken = "&pageToken=";
+            $reload = true;
 
-        $youtube = Http::get("{$apiUrl}{$part}{$channelId}{$key}{$maxResults}{$nextPageToken}");
-        $youtubeList = $youtube->json();
-
-        if (array_key_exists('items',$youtubeList)) {
-            if (array_key_exists('nextPageToken',$youtubeList)) {
-                $nextPageToken = "&pageToken={$youtubeList['nextPageToken']}";
-            } else {
-                $reload = false;
-            }
-
-            foreach ($youtubeList['items'] as $key => $value) {
-                if (array_key_exists('videoId',$value['id'])) {
-                    Video::create([
-                        'gambar'        => $value['snippet']['thumbnails']['high']['url'],
-                        'video_id'      => $value['id']['videoId'],
-                        'caption'       => $value['snippet']['title'],
-                        'published_at'  => date('Y-m-d H:i:s',strtotime($value['snippet']['publishedAt'])),
-                    ]);
-                }
-            }
-
-            while ($reload) {
+            do {
                 $youtube = Http::get("{$apiUrl}{$part}{$channelId}{$key}{$maxResults}{$nextPageToken}");
                 $youtubeList = $youtube->json();
-                if (array_key_exists('nextPageToken',$youtubeList)) {
-                    $nextPageToken = "&pageToken={$youtubeList['nextPageToken']}";
-                } else {
-                    $reload = false;
-                }
+                if (array_key_exists('items',$youtubeList)) {
+                    if (array_key_exists('nextPageToken',$youtubeList)) {
+                        $nextPageToken = "&pageToken={$youtubeList['nextPageToken']}";
+                    } else {
+                        $reload = false;
+                    }
 
-                foreach ($youtubeList['items'] as $key => $value) {
-                    if (array_key_exists('videoId',$value['id'])) {
-                        Video::create([
-                            'gambar'        => $value['snippet']['thumbnails']['high']['url'],
-                            'video_id'      => $value['id']['videoId'],
-                            'caption'       => $value['snippet']['title'],
-                            'published_at'  => date('Y-m-d H:i:s',strtotime($value['snippet']['publishedAt'])),
-                        ]);
+                    foreach ($youtubeList['items'] as $key => $value) {
+                        if (array_key_exists('videoId',$value['id'])) {
+                            Video::create([
+                                'gambar'        => $value['snippet']['thumbnails']['high']['url'],
+                                'video_id'      => $value['id']['videoId'],
+                                'caption'       => $value['snippet']['title'],
+                                'published_at'  => date('Y-m-d H:i:s',strtotime($value['snippet']['publishedAt'])),
+                            ]);
+                        }
+                    }
+                } else {
+                    if (Video::count() > 0) {
+                        $reload = false;
+                    } else {
+                        return back()->with('api_error', $youtubeList['error']['message']);
                     }
                 }
-            }
-
+            } while ($reload);
             return back()->with('success', 'Video berhasil disinkronisasi');
         } else {
-            return back()->with('api_error', $youtubeList['error']['message']);
+            return back()->with('error', 'Harap memasukkan API KEY pada .env');
         }
     }
